extract git stuff in own class
